refactor(phylopic): move names to their own files
