+++ Phylopic/src/Phylopic.jl
@@ -6,7 +6,7 @@ using UUIDs
 
 const api = "https://api.phylopic.org/"
 
-include(joinpath(pwd(), "src", "ping.jl"))
+include("ping.jl")
 
 # We do a first ping to fail ASAP if the API is not responsive
 @assert isnothing(Phylopic.ping())
@@ -18,12 +18,9 @@ const buildnumber = Phylopic.build()
 
 # the autocomplete endpoint is meant to give an overview of possible names starting from
 # a stem - this is not necessarilly going to give all of the names, and I am not sure why
-include(joinpath(pwd(), "src", "autocomplete.jl"))
+include("autocomplete.jl")
 
-# the images endpoint does a lot of different things, which is very confusing but that's the
-# way the API is built. When given search parameters, it gives back a range of pages. When
-# given a page, it gives a list of items. When given an UUID, it gives the links to all of
-# the versions of an image.
-include(joinpath(pwd(), "src", "images.jl"))
+include("names.jl")
+include("images.jl")
 
-end # module Phylopic
+end # module hylopic
